Batch workload generation + automated runs
Usability improvement - allows a user to define a
"workload matrix", the rows of which define a COSbench
workload and columns can be configured with range of
object sizes, range of number of containers and a series
of Read/Write/Delete mixes.

Each workload row can be repeated varying the number
of COSbench workers (clients).

This enhancement should help with scalability analyses
of cloud object storage systems.

Signed-off-by: Gitesh Nandre <gitesh.e.nandre@intel.com>
Signed-off-by: Tushar Gohad <tushar.gohad@intel.com>